requirements.txt
flask
fastapi
uvicorn

sqlalchemy

sentence-transformers
langchain

pinecone-client
weaviate-client  
faiss-cpu  

pandas
numpy

requests

gunicorn  

loguru

scikit-learn
transformers 
openai  

anthropic
python-dotenv
transformers
torch
accelerate
